+++ packages/connect-react/src/hooks/form-context.tsx
@@ -44,6 +44,7 @@ export type FormContext<T extends ConfigurableProps> = {
   /** @deprecated Use externalUserId instead */
   userId: string;
   enableDebugging?: boolean;
+  oauthAppId?: string;
 };
 
 export const skippablePropTypes = [
@@ -80,7 +81,7 @@ export const FormContextProvider = <T extends ConfigurableProps>({
   const id = useId();
 
   const {
-    component, configuredProps: __configuredProps, propNames, externalUserId, userId, sdkResponse, enableDebugging,
+    component, configuredProps: __configuredProps, propNames, externalUserId, userId, sdkResponse, enableDebugging, oauthAppId,
   } = formProps;
 
   // Resolve user ID with deprecation warning
@@ -593,6 +594,7 @@ export const FormContextProvider = <T extends ConfigurableProps>({
     submitting,
     sdkErrors,
     enableDebugging,
+    oauthAppId,
   };
   return <FormContext.Provider value={value}>{children}</FormContext.Provider>;
 };